Range buttons component split
