+++ src/features/charts/components/index.ts
@@ -1,2 +1,3 @@
 export * from './graph'
-export * from './rangeButtons'
+export * from './RangeComponent'
+export * from './RangeButton'
